Safely create resolvers for apollo client
