+++ packages/venia-ui/lib/components/CartPage/PriceAdjustments/GiftOptions/giftOptions.js
@@ -1,12 +1,11 @@
 import React from 'react';
-import gql from 'graphql-tag';
 
 import useGiftOptions from '@magento/peregrine/lib/talons/CartPage/GiftOptions/useGiftOptions';
 
 import Checkbox from '../../../Checkbox';
 import TextArea from '../../../TextArea';
 import { mergeClasses } from '../../../../classify';
-
+import GiftOptionsOperations from './giftOptions.gql';
 import defaultClasses from './giftOptions.css';
 
 const GiftOptions = props => {
@@ -18,8 +17,7 @@ const GiftOptions = props => {
         toggleIncludePrintedCardFlag,
         updateGiftMessage
     } = useGiftOptions({
-        getGiftOptionsQuery: GIFT_OPTIONS_QUERY,
-        saveGiftOptionsQuery: SET_GIFT_OPTIONS_QUERY
+        ...GiftOptionsOperations
     });
 
     const classes = mergeClasses(defaultClasses, props.classes);
@@ -59,103 +57,6 @@ const GiftOptions = props => {
             </ul>
         </div>
     );
-};
-
-/**
- * Local query. GQL support is not available as of today.
- *
- * Once available, we can change the query to match the schema.
- */
-const GIFT_OPTIONS_QUERY = gql`
-    query getGiftOptions($cart_id: String) {
-        gift_options(cart_id: $cart_id) @client @connection(key: "Cart") {
-            include_gift_receipt
-            include_printed_card
-            gift_message
-        }
-    }
-`;
-
-/**
- * Local mutation. GQL support is not available as of today.
- *
- * Once available, we can change the mutation to match the schema.
- */
-const SET_GIFT_OPTIONS_QUERY = gql`
-    mutation setGiftOptions(
-        $cart_id: String
-        $include_gift_receipt: Boolean
-        $include_printed_card: Boolean
-        $gift_message: String
-    ) {
-        set_gift_options(
-            cart_id: $cart_id
-            include_gift_receipt: $include_gift_receipt
-            include_printed_card: $include_printed_card
-            gift_message: $gift_message
-        ) @client @connection(key: "set_gift_options")
-    }
-`;
-
-export const giftOptionsResolvers = {
-    Query: {
-        gift_options: (_, { cart_id }, { cache, getCacheKey }) => {
-            /**
-             * This is how the `cacheKeyFromType` saves the
-             * cart data in the `InMemoryCache`.
-             */
-            const cartIdInCache = getCacheKey({
-                __typename: 'Cart',
-                id: cart_id
-            });
-
-            const { include_gift_receipt, include_printed_card, gift_message } =
-                cache.data.data[cartIdInCache] || {};
-
-            return {
-                __typename: 'Cart',
-                include_gift_receipt,
-                include_printed_card,
-                gift_message
-            };
-        }
-    },
-    Mutation: {
-        set_gift_options: (
-            _,
-            {
-                cart_id,
-                include_gift_receipt = false,
-                include_printed_card = false,
-                gift_message = ''
-            },
-            { cache }
-        ) => {
-            cache.writeQuery({
-                query: GIFT_OPTIONS_QUERY,
-                variables: {
-                    cart_id
-                },
-                data: {
-                    gift_options: {
-                        include_gift_receipt,
-                        include_printed_card,
-                        gift_message,
-                        id: cart_id,
-                        __typename: 'Cart'
-                    }
-                }
-            });
-
-            /**
-             * We do not return anything on purpose.
-             * Returning something here will update the component
-             * and it will cause the text area to re render and
-             * create a snappy effect.
-             */
-            return null;
-        }
-    }
 };
 
 export default GiftOptions;
