+++ packages/venia-ui/lib/components/CartPage/PriceAdjustments/GiftOptions/index.js
@@ -1,1 +1,1 @@
-export { default, giftOptionsResolvers } from './giftOptions';
+export { default } from './giftOptions';
