+++ packages/venia-ui/lib/components/CartPage/PriceAdjustments/index.js
@@ -1,3 +1,1 @@
 export { default } from './priceAdjustments';
-
-export { giftOptionsResolvers } from './GiftOptions';
