+++ packages/venia-ui/lib/resolvers/resolvers.js
@@ -1,14 +1,28 @@
-import { giftOptionsResolvers } from '../components/CartPage/PriceAdjustments';
+import { giftOptionsResolvers } from '../components/CartPage/PriceAdjustments/GiftOptions/giftOptions.gql';
 import { checkoutPageResolvers } from '../components/CheckoutPage/checkoutPage.gql';
 
-export default {
-    Query: {
-        ...giftOptionsResolvers.Query
-    },
-    Mutation: {
-        ...giftOptionsResolvers.Mutation
-    },
-    // TODO: Is this safe? Apparently you can have `Type` resolvers. Not sure if
-    // checkoutPageResolvers.Query would overwrite giftOptionsResolvers.
-    ...checkoutPageResolvers
-};
+/**
+ * Type resolvers are merged by the client so spread each resolver into a
+ * separate object.
+ *
+ * NOTE: Be careful not to overwrite type properties. For example, suppose two
+ * resolvers are spread into the array resulting in the following. "foo" will be
+ * overwritten while "bar" and "baz" will not be.
+ *
+ * [
+ *   { // From Component A resolvers
+ *     Query: {
+ *       foo: () => { return 1; }
+ *       bar: () => { return 2; }
+ *     }
+ *   },
+ *   { // From Component B resolvers
+ *     Query: {
+ *       foo: () => { return 3; }
+ *       baz: () => { return 4; }
+ *     }
+ *   }
+ * ]
+ */
+
+export default [{ ...giftOptionsResolvers }, { ...checkoutPageResolvers }];
